Initial replacement of Facebook SDK with custom code
We use WebView to handle the custom OAuth flow on which end we
obtain the access token. We then use AsyncHttp library to perform
REST queries on the Facebook API.

The need to use custom code is that we need a single way to obtain
user session cookies that we can then re-use when scraping the
Facebook website for the birthday URL. The Facebook SDK with
its native flow won't give us the cookies and having two login
screens just to get the cookies feels weird.

This patch is only missing handling for expired and invalid tokens,
ideally in the background without too any user interaction.