Persist past events in the calendar
Facebook does not list events once they pass, but it's desirable that we
keep them in the calendar. For that reason we don't ever remove them,
assuming that users would not remove themselves from past events (only
from future events, and that we still handle as expected).

For birthdays we set the year to current year (year of sync) minus one
and set their recurrence to yearly, so that they don't vanish the day
after.