Overhaul token handling
Store  UID and key for iCal and API access token as auth tokens for
the account, the iCal URL then fully reconstructed using the parameters
instead of clumsily replacing schema and path.

Also makes tokens to be properly retrieved and re-requested from user
when they are invalidated.

This should fix the issue with events or birthdays not being retrieved.