Bind WebView cookies to Account (fixes #18)
WebView cookies are now stored as part of Account userData and removed from the
WebView cookie store once the authentication is over. This way removing the account
from the app actually resets the cookies and adding a new account shows the
Facebook login prompt.

This also enables support for multiple accounts, although we don't have a proper
UI for that yet.