Redesign main activity to handle multiple accounts, add About page
I'm too lazy to split the commit at this point, but it ports the main
activity to use a navigation drawer for menu and the main screen content
is just account list using a RecyclerView. Also adds About activity that
with license and version informations and links to our Privacy Policy and
Play Store.