Implement custom ContentProvider to handle preferences
We need to share our preferences settings between what are effectively two
processes: the GUI app and the Sync adapter. The SharedPreferences are not
suitable for this as the Sync adapter does not see changes done to the
SharedPreferences in the GUI app.

To fix that we implement a custom ContentProvider backed with SQLite database
to safely and efficiently synchronize configuration preferences between
the two processes.

To integrate the custom ContentProvider into the PreferencesActivity we
need to use PreferencesDateStore, which was introduced in API 27, so in order
to support that we had to pull android.support modules for Preferences and
Fragments, hence this change is a little bit more invasive than expected.