Roll back the sharedUserId change in Manifests
Apparently adding sharedUserId makes it impossible for users to update
the app if they already have it installed and the only way around is
to manually uninstall it and install it again.

Since this thing is only needed for the instrumented test to get the
right permissions to access calendars, we can remove it for now and
try to find some other way to make the test work without sharedUserId.

Man, Android really sucks sometimes...