Fix iCal link retrieval due to Facebook changes
Looks like Facebook has decided to remove the 'Upcoming birthdays'
link from the /events page, instead the link is now displayed in
an async dialog when user clicks 'Export event' in any event.

Clicking the button does an ajax call, which we attempt to replay
again a custom event I've created and then try to parse the key
out from the returned text. I can't image this getting any more
fragile, but it's the only way we have right now.